requirements.txt
# Dépendances communes
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
implicit==0.6.2
mlflow==2.11.3
psycopg2-binary==2.9.9
sqlalchemy==2.0.30
# API & UI
fastapi==0.111.0
uvicorn==0.29.0
pydantic==2.7.0
streamlit==1.35.0
plotly==5.22.0
requests==2.32.2
Pillow==10.3.0
